Refactor directory structure
+ Improve database query
