+++ src/database/provider.ts
@@ -1,0 +1,9 @@
+export interface Provider {
+  id: number;
+  baseURL?: string;
+  apiKey?: string;
+}
+
+export interface ProviderWithCount extends Provider {
+  modelCount: number;
+}
