--- src/lib/utils.types.ts
@@ -1,4 +1,0 @@
-export enum UiToggleState {
-  OPEN,
-  CLOSE,
-}
