+++ src/pages/about/index.tsx
@@ -1,4 +1,3 @@
-import Drawer from "@components/drawer";
 import { getName, getVersion } from "@tauri-apps/api/app";
 import { useEffect, useState } from "react";
 
@@ -16,11 +15,6 @@ export default function About() {
   return (
     <div>
       {/* navbar */}
-      <div className="navbar bg-base-100 flex-none px-6 flex sticky top-0 z-10">
-        <div className="navbar-start me-6">
-          <Drawer />
-        </div>
-      </div>
       {/* title */}
       <div className="px-7 py-4 space-y-4 max-w-md">
         <img src="/elelem.png" alt="icon" className="h-16 w-16" />
