+++ src/pages/prompt/add.modal.tsx
@@ -1,8 +1,7 @@
 import SubmitButton from "@components/submit-button";
 import { yupResolver } from "@hookform/resolvers/yup";
-import db from "@lib/database";
-import { toggleModal } from "@lib/utils";
-import { UiToggleState } from "@lib/utils.types";
+import db from "@database/config";
+import { toggleModal, UiToggleState } from "@utils/toggle";
 import { IconX } from "@tabler/icons-react";
 import { useForm } from "react-hook-form";
 import * as yup from "yup";
@@ -10,12 +9,10 @@ import * as yup from "yup";
 export const AddPromptModalId = "addPromptModal";
 
 export default function AddPromptModal() {
-  const schema = yup
-    .object()
-    .shape({
-      title: yup.string().label("Title").required(),
-      prompt: yup.string().label("Prompt").required(),
-    });
+  const schema = yup.object().shape({
+    title: yup.string().label("Title").required(),
+    prompt: yup.string().label("Prompt").required(),
+  });
 
   const {
     reset,
