+++ src/pages/provider/add.modal.tsx
@@ -1,14 +1,13 @@
 import SubmitButton from "@components/submit-button";
 import { yupResolver } from "@hookform/resolvers/yup";
-import db from "@lib/database";
-import { fetchModels } from "@lib/model";
-import { toggleModal } from "@lib/utils";
-import { UiToggleState } from "@lib/utils.types";
+import db from "@database/config";
+import { toggleModal, UiToggleState } from "@utils/toggle";
 import { IconCircleX, IconX } from "@tabler/icons-react";
 import { isEmpty } from "radash";
 import { useState } from "react";
 import { useForm } from "react-hook-form";
 import * as yup from "yup";
+import { fetchModelList } from "@utils/conversation";
 
 export const AddProviderModalId = "addProviderModal";
 
@@ -41,7 +40,7 @@ export default function AddProviderModal() {
     // get model list
     let modelIds: string[] = [];
     try {
-      modelIds = await fetchModels(baseURL, data.apiKey);
+      modelIds = await fetchModelList(baseURL, data.apiKey);
     } catch (_error) {
       setError("Failed to connect");
       return;
