+++ src/stores/prompt.ts
